# openapi.json
openapi: 3.0.3
info:
  title: My Dear Partner API
  version: 1.0.0
  description: API endpoints documentation for the My Dear Partner platform.
paths:
...
  /api/v1/admin/accounts/{account_id}/:
    get:
      operationId: v1_admin_accounts_retrieve_2
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
    patch:
      operationId: v1_admin_accounts_partial_update
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
    delete:
      operationId: v1_admin_accounts_destroy
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      responses:
        '204':
          description: No response body
...
  /api/v1/super-admin/accounts/{account_type}/{account_id}/:
    get:
      operationId: v1_super_admin_accounts_retrieve_2
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: account_type
        schema:
          type: string
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
    patch:
      operationId: v1_super_admin_accounts_partial_update
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: account_type
        schema:
          type: string
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
    delete:
      operationId: v1_super_admin_accounts_destroy
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: account_type
        schema:
          type: string
        required: true
      tags:
      - v1
      responses:
        '204':
          description: No response body
  /api/v1/super-admin/accounts/{account_type}/{account_id}/{action}/:
    post:
      operationId: v1_super_admin_accounts_create_2
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: account_type
        schema:
          type: string
        required: true
      - in: path
        name: action
        schema:
          type: string
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
...
  /api/v1/super-admin/admins/{account_id}/:
    get:
      operationId: v1_super_admin_admins_retrieve_2
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
    patch:
      operationId: v1_super_admin_admins_partial_update
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      responses:
        '200':
          description: No response body
    delete:
      operationId: v1_super_admin_admins_destroy
      parameters:
      - in: path
        name: account_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      responses:
        '204':
          description: No response body
...